criado-demais-tabelas

+++ src/main/java/br/com/wanessacamara/estoque/service/ProdutoService.java
@@ -4,6 +4,7 @@ import br.com.wanessacamara.estoque.model.Produto;
 import br.com.wanessacamara.estoque.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,11 +22,18 @@ public class ProdutoService {
     public Produto criaProduto(Produto produto) {
         return repository.save(produto);
     }
-
+    @Transactional
     public void deletaProduto(Long id) {
-        repository.deleteById(id);
+        Optional<Produto> optionalProduto = repository.findById(id);
+
+        if (optionalProduto.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new RuntimeException("Produto não encontrado com o ID: " + id);
+        }
     }
 
+    @Transactional
     public Produto atualizaProduto(Long id, Produto produtoAtualizado) {
         Optional<Produto> optionalProduto = repository.findById(id);
